- Adds 'filter-standard' which (to run easily)   required unpicking the parser from utils   into a new parser in std_utils
